requirements.txt
mysql-connector-python==8.1.0
flask==2.3.3
python-dotenv==1.0.0